Lineups are now stored dating back to 2020-2021.
Each week, uploaded csv will tally points,
configure the standings.

Then it will look at each teams lineup. and Then
take in the total drafted player table.

Going to next make a team page, which can be
a modal pop up.  Same with player page.

All stats as of now can be dkdfs oriented.

Goal in the future will be to bring in there
actual stats on the year.

And make this similar to nbaportal.